Move execution-related stuff to separate class
